requirements.txt
# Core dependencies
python-dotenv>=1.0.0
PyYAML>=6.0

# PDF processing
pdfplumber>=0.10.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Video processing
moviepy>=1.0.3

# Web scraping for brand extraction
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# AI/LLM (OpenAI)
openai>=1.3.0

# CLI interface
click>=8.1.0

# Utilities
colorama>=0.4.6
